# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"MovimentacaoResponseDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"clienteId": {
						"type": "number",
						"format": "double"
					},
					"coxinhaId": {
						"type": "number",
						"format": "double"
					},
					"dataHora": {
						"type": "string",
						"format": "date-time"
					},
					"valorPago": {
						"type": "number",
						"format": "double"
					},
					"troco": {
						"type": "number",
						"format": "double"
					},
					"tipoSabor": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"clienteId",
					"coxinhaId",
					"dataHora",
					"valorPago",
					"troco",
					"tipoSabor",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BasicResponseDTO_MovimentacaoResponseDTO_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"object": {
						"$ref": "#/components/schemas/MovimentacaoResponseDTO"
					}
				},
				"required": [
					"message",
					"object"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Movimentacao": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"BasicResponseDTO_Movimentacao_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"object": {
						"$ref": "#/components/schemas/Movimentacao"
					}
				},
				"required": [
					"message",
					"object"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BasicResponseDTO_boolean_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"object": {
						"type": "boolean"
					}
				},
				"required": [
					"message",
					"object"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BasicResponseDTO_MovimentacaoResponseDTO-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"object": {
						"items": {
							"$ref": "#/components/schemas/MovimentacaoResponseDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"message",
					"object"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CoxinhaResponseDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"sabor": {
						"type": "string"
					},
					"preco": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"sabor",
					"preco"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BasicResponseDTO_CoxinhaResponseDTO-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"object": {
						"items": {
							"$ref": "#/components/schemas/CoxinhaResponseDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"message",
					"object"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ClienteResponseDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"nome": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"saldo": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"nome",
					"email",
					"saldo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BasicResponseDTO_ClienteResponseDTO_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"object": {
						"$ref": "#/components/schemas/ClienteResponseDTO"
					}
				},
				"required": [
					"message",
					"object"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ClienteRequestDTO": {
				"properties": {
					"nome": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"senha": {
						"type": "string"
					}
				},
				"required": [
					"nome",
					"email",
					"senha"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "API Documentacao",
		"version": "1.0.0",
		"description": "Documentacao da API usando Swagger"
	},
	"paths": {
		"/movimentacoes": {
			"post": {
				"operationId": "InserirMovimentacao",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_MovimentacaoResponseDTO_"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_Movimentacao_"
								}
							}
						}
					}
				},
				"tags": [
					"Movimentações"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"valorInserido": {
										"type": "number",
										"format": "double"
									},
									"coxinhaId": {
										"type": "number",
										"format": "double"
									},
									"clienteId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"valorInserido",
									"coxinhaId",
									"clienteId"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListaMovimentacoes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_MovimentacaoResponseDTO-Array_"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_MovimentacaoResponseDTO-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Movimentações"
				],
				"security": [],
				"parameters": []
			}
		},
		"/movimentacoes/estornar": {
			"post": {
				"operationId": "EstornarMovimentacao",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_boolean_"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_boolean_"
								}
							}
						}
					}
				},
				"tags": [
					"Movimentações"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"id": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/movimentacoes/trocar-sabor": {
			"post": {
				"operationId": "TrocarSabor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_boolean_"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_boolean_"
								}
							}
						}
					}
				},
				"tags": [
					"Movimentações"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"novaCoxinhaId": {
										"type": "number",
										"format": "double"
									},
									"movimentacaoId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"novaCoxinhaId",
									"movimentacaoId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/catalogo/coxinha": {
			"get": {
				"operationId": "ListaCoxinhas",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_CoxinhaResponseDTO-Array_"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_CoxinhaResponseDTO-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Coxinha"
				],
				"security": [],
				"parameters": []
			}
		},
		"/login/cliente": {
			"post": {
				"operationId": "InserirCliente",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_ClienteResponseDTO_"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_ClienteResponseDTO_"
								}
							}
						}
					}
				},
				"tags": [
					"Cliente"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ClienteRequestDTO"
							}
						}
					}
				}
			}
		},
		"/login/cliente/login": {
			"post": {
				"operationId": "FazerLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_ClienteResponseDTO_"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_ClienteResponseDTO_"
								}
							}
						}
					}
				},
				"tags": [
					"Cliente"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"senha": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"senha",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/login/cliente/saldo": {
			"put": {
				"operationId": "AtualizaSaldo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_boolean_"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_ClienteResponseDTO_"
								}
							}
						}
					}
				},
				"tags": [
					"Cliente"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"saldo": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"saldo"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/login/cliente/{id}": {
			"get": {
				"operationId": "BuscarClientePorId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_ClienteResponseDTO_"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDTO_ClienteResponseDTO_"
								}
							}
						}
					}
				},
				"tags": [
					"Cliente"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3090/api",
			"description": "Servidor de Desenvolvimento"
		}
	]
}